refactor: use word loads in note script, add byte-swap doc, relocate MASM unit tests
- Note script: replace 14 individual mem_load instructions with 4 idiomatic
  padw + mem_loadw_le word loads (matching b2agg.masm pattern)
- message_note.rs: document LE byte-swap convention in build_note_storage
- Move test_from_account_id_matches_rust to solidity_miden_address_conversion.rs
- Move compute_leaf_value message-type test to leaf_utils.rs
