+++ crates/miden-testing/tests/agglayer/bridge_message.rs
@@ -516,305 +516,6 @@ async fn bridge_message_non_target_cannot_consume() -> anyhow::Result<()> {
     Ok(())
 }
 
-/// Standalone MASM unit test that mimics how `bridge_message` stores leaf data to memory,
-/// calls `compute_leaf_value`, and compares the hash with an independently computed Keccak256.
-///
-/// This bypasses the full transaction pipeline and directly exercises the MASM memory layout
-/// and hash computation to diagnose leaf hash mismatches.
-#[tokio::test]
-async fn bridge_message_leaf_hash_masm_unit_test() -> anyhow::Result<()> {
-    use alloc::sync::Arc;
-    use alloc::vec::Vec;
-
-    use miden_agglayer::agglayer_library;
-    use miden_assembly::{Assembler, DefaultSourceManager};
-    use miden_core_lib::CoreLibrary;
-    use miden_crypto::hash::keccak::Keccak256;
-    use miden_processor::utils::{bytes_to_packed_u32_elements, packed_u32_elements_to_bytes};
-    use miden_protocol::Felt;
-
-    use super::test_utils::execute_program_with_default_host;
-
-    // --- Rust: build the 113-byte packed representation ---
-    let origin_address_bytes = [0u8; 20];
-    let destination_address_bytes = [0u8; 20];
-    let metadata_hash_bytes = [0x42u8; 32];
-
-    let mut packed = Vec::with_capacity(113);
-    packed.push(1u8); // leafType = 1 (message)
-    packed.extend_from_slice(&77u32.to_be_bytes()); // originNetwork = MIDEN_NETWORK_ID
-    packed.extend_from_slice(&origin_address_bytes); // 20 bytes
-    packed.extend_from_slice(&1u32.to_be_bytes()); // destinationNetwork = 1
-    packed.extend_from_slice(&destination_address_bytes); // 20 bytes
-    packed.extend_from_slice(&[0u8; 32]); // amount = 0
-    packed.extend_from_slice(&metadata_hash_bytes); // 32 bytes
-    assert_eq!(packed.len(), 113);
-
-    let expected_hash = Keccak256::hash(&packed);
-    let expected_felts: Vec<Felt> = bytes_to_packed_u32_elements(expected_hash.as_bytes());
-
-    // --- Compute the u32 felt values for each field as bridge_message would store them ---
-
-    // leafType = 1, byte-swapped: swap_u32_bytes(1) = 0x01000000 = 16777216
-    let leaf_type_swapped: u32 = 16777216;
-
-    // originNetwork = 77, byte-swapped: swap_u32_bytes(77) = 0x4D000000 = 1291845632
-    let origin_network_swapped: u32 = 1291845632;
-
-    // originAddress = 5 zeros (all-zero address)
-    // Each limb is bytes_to_packed_u32_elements([0,0,0,0]) = 0
-    let origin_addr_felts = [0u32; 5];
-
-    // destinationNetwork = 1, byte-swapped: swap_u32_bytes(1) = 0x01000000 = 16777216
-    let dest_network_swapped: u32 = 16777216;
-
-    // destinationAddress = 5 zeros
-    let dest_addr_felts = [0u32; 5];
-
-    // amount = 8 zeros
-    // metadataHash: [0x42; 32] as packed u32 LE felts
-    // [0x42, 0x42, 0x42, 0x42] -> u32::from_le_bytes = 0x42424242 = 1111638594
-    let metadata_felts: Vec<Felt> = bytes_to_packed_u32_elements(&metadata_hash_bytes);
-    assert_eq!(metadata_felts.len(), 8);
-
-    // --- Build MASM program ---
-    // Memory layout (start_ptr = 0):
-    //   addr 0:     leafType (byte-swapped)
-    //   addr 1:     originNetwork (byte-swapped)
-    //   addr 2-6:   originAddress (5 felts)
-    //   addr 7:     destinationNetwork (byte-swapped)
-    //   addr 8-12:  destinationAddress (5 felts)
-    //   addr 13-20: amount (8 felts, all zero)
-    //   addr 21-28: metadataHash (8 felts)
-    //   addr 29-31: padding (3 felts, all zero)
-
-    let mh: Vec<u64> = metadata_felts.iter().map(|f| f.as_canonical_u64()).collect();
-
-    let source = format!(
-        r#"
-            use miden::core::sys
-            use agglayer::bridge::leaf_utils
-
-            begin
-                # Store leafType = 1 (byte-swapped)
-                push.{leaf_type_swapped} push.0 mem_store
-
-                # Store originNetwork = 77 (byte-swapped)
-                push.{origin_network_swapped} push.1 mem_store
-
-                # Store originAddress = 5 zeros
-                push.{oa0} push.2 mem_store
-                push.{oa1} push.3 mem_store
-                push.{oa2} push.4 mem_store
-                push.{oa3} push.5 mem_store
-                push.{oa4} push.6 mem_store
-
-                # Store destinationNetwork = 1 (byte-swapped)
-                push.{dest_network_swapped} push.7 mem_store
-
-                # Store destinationAddress = 5 zeros
-                push.{da0} push.8 mem_store
-                push.{da1} push.9 mem_store
-                push.{da2} push.10 mem_store
-                push.{da3} push.11 mem_store
-                push.{da4} push.12 mem_store
-
-                # Store amount = 8 zeros
-                push.0 push.13 mem_store
-                push.0 push.14 mem_store
-                push.0 push.15 mem_store
-                push.0 push.16 mem_store
-                push.0 push.17 mem_store
-                push.0 push.18 mem_store
-                push.0 push.19 mem_store
-                push.0 push.20 mem_store
-
-                # Store metadataHash = 8 felts
-                push.{mh0} push.21 mem_store
-                push.{mh1} push.22 mem_store
-                push.{mh2} push.23 mem_store
-                push.{mh3} push.24 mem_store
-                push.{mh4} push.25 mem_store
-                push.{mh5} push.26 mem_store
-                push.{mh6} push.27 mem_store
-                push.{mh7} push.28 mem_store
-
-                # Store padding = 3 zeros
-                push.0 push.29 mem_store
-                push.0 push.30 mem_store
-                push.0 push.31 mem_store
-
-                # Compute leaf value (start_ptr = 0)
-                push.0
-                exec.leaf_utils::compute_leaf_value
-
-                exec.sys::truncate_stack
-            end
-        "#,
-        leaf_type_swapped = leaf_type_swapped,
-        origin_network_swapped = origin_network_swapped,
-        oa0 = origin_addr_felts[0],
-        oa1 = origin_addr_felts[1],
-        oa2 = origin_addr_felts[2],
-        oa3 = origin_addr_felts[3],
-        oa4 = origin_addr_felts[4],
-        dest_network_swapped = dest_network_swapped,
-        da0 = dest_addr_felts[0],
-        da1 = dest_addr_felts[1],
-        da2 = dest_addr_felts[2],
-        da3 = dest_addr_felts[3],
-        da4 = dest_addr_felts[4],
-        mh0 = mh[0],
-        mh1 = mh[1],
-        mh2 = mh[2],
-        mh3 = mh[3],
-        mh4 = mh[4],
-        mh5 = mh[5],
-        mh6 = mh[6],
-        mh7 = mh[7],
-    );
-
-    let agglayer_lib = agglayer_library();
-    let program = Assembler::new(Arc::new(DefaultSourceManager::default()))
-        .with_dynamic_library(CoreLibrary::default())
-        .unwrap()
-        .with_dynamic_library(agglayer_lib)
-        .unwrap()
-        .assemble_program(&source)
-        .unwrap();
-
-    let exec_output = execute_program_with_default_host(program, None).await?;
-
-    // Note: pack_leaf_data modifies memory in-place, so post-execution memory contains the
-    // packed result, not the original stored values. We focus on the hash comparison.
-    let ctx = miden_processor::ContextId::root();
-
-    // --- Compare hash output ---
-    let computed_felts: Vec<Felt> = exec_output.stack[0..8].to_vec();
-    let computed_u32s: Vec<u32> =
-        computed_felts.iter().map(|f| f.as_canonical_u64() as u32).collect();
-    let expected_u32s: Vec<u32> =
-        expected_felts.iter().map(|f| f.as_canonical_u64() as u32).collect();
-
-    if computed_u32s != expected_u32s {
-        // Diagnostic: show both as bytes
-        let computed_bytes = packed_u32_elements_to_bytes(&computed_felts);
-        let expected_bytes = packed_u32_elements_to_bytes(&expected_felts);
-        eprintln!("MASM hash (felts):     {:?}", computed_u32s);
-        eprintln!("Expected hash (felts): {:?}", expected_u32s);
-        eprintln!("MASM hash (bytes):     {:02x?}", computed_bytes);
-        eprintln!("Expected hash (bytes): {:02x?}", expected_bytes);
-
-        // Also dump the packed memory for diagnosis (after pack_leaf_data ran)
-        let packed_mem: Vec<u32> = (0..29u32)
-            .map(|addr| {
-                exec_output
-                    .memory
-                    .read_element(ctx, Felt::from(addr))
-                    .map(|f| f.as_canonical_u64() as u32)
-                    .unwrap_or(0)
-            })
-            .collect();
-        eprintln!("Packed memory (29 u32 felts): {:08x?}", packed_mem);
-
-        // Convert packed memory to bytes for comparison with expected packed bytes
-        let packed_mem_felts: Vec<Felt> = (0..29u32)
-            .map(|addr| {
-                exec_output.memory.read_element(ctx, Felt::from(addr)).unwrap_or(Felt::ZERO)
-            })
-            .collect();
-        let packed_mem_bytes = packed_u32_elements_to_bytes(&packed_mem_felts);
-        eprintln!(
-            "Packed memory bytes (first 113): {:02x?}",
-            &packed_mem_bytes[..113.min(packed_mem_bytes.len())]
-        );
-        eprintln!("Expected packed bytes:           {:02x?}", &packed);
-    }
-
-    assert_eq!(
-        computed_u32s, expected_u32s,
-        "MASM compute_leaf_value output should match Rust Keccak256 of packed leaf data"
-    );
-
-    Ok(())
-}
-
-/// MASM unit test: verifies that `eth_address::from_account_id` produces the same 5 felts
-/// as the Rust `EthEmbeddedAccountId::from_account_id().to_elements()`.
-///
-/// This isolates the origin-address conversion from the full bridge pipeline to diagnose
-/// leaf hash mismatches.
-#[tokio::test]
-async fn test_from_account_id_matches_rust() -> anyhow::Result<()> {
-    use miden_agglayer::EthEmbeddedAccountId;
-    use miden_protocol::account::{AccountId, AccountIdVersion, AccountType};
-
-    use super::test_utils::execute_masm_script;
-
-    // Pick a deterministic AccountId
-    let account_id = AccountId::dummy([42; 15], AccountIdVersion::Version1, AccountType::Public);
-
-    // Rust: compute expected 5 felts
-    let expected_elements = EthEmbeddedAccountId::from_account_id(account_id).to_elements();
-    assert_eq!(expected_elements.len(), 5, "EthAddress should have 5 elements");
-
-    // Get the raw felt values for MASM
-    let prefix_val = account_id.prefix().as_u64();
-    let suffix_val = account_id.suffix().as_canonical_u64();
-
-    eprintln!("AccountId: prefix={prefix_val}, suffix={suffix_val}");
-    eprintln!(
-        "Expected elements (u32): {:?}",
-        expected_elements
-            .iter()
-            .map(|f| f.as_canonical_u64() as u32)
-            .collect::<Vec<_>>()
-    );
-
-    // MASM program: push prefix then suffix so stack is [suffix, prefix]
-    // (push.A push.B gives [B, A] on stack, B on top)
-    // from_account_id expects [suffix, prefix] with suffix on top
-    let source = format!(
-        r#"
-            use miden::core::sys
-            use agglayer::common::eth_address
-
-            begin
-                push.{prefix_val} push.{suffix_val}
-                # => [suffix, prefix]  (suffix on top)
-                exec.eth_address::from_account_id
-                # => [limb0, limb1, limb2, limb3, limb4]
-                exec.sys::truncate_stack
-            end
-        "#,
-        prefix_val = prefix_val,
-        suffix_val = suffix_val,
-    );
-
-    let exec_output = execute_masm_script(&source).await?;
-
-    // Read the 5 felts from the stack (limb0 on top = stack[0])
-    let actual: Vec<u64> = exec_output.stack[0..5].iter().map(|f| f.as_canonical_u64()).collect();
-
-    let expected: Vec<u64> = expected_elements.iter().map(|f| f.as_canonical_u64()).collect();
-
-    eprintln!("MASM actual   (u64): {:?}", actual);
-    eprintln!("Rust expected (u64): {:?}", expected);
-
-    // Also show as u32 for easier reading
-    let actual_u32: Vec<u32> = actual.iter().map(|v| *v as u32).collect();
-    let expected_u32: Vec<u32> = expected.iter().map(|v| *v as u32).collect();
-    eprintln!("MASM actual   (u32): {:?}", actual_u32);
-    eprintln!("Rust expected (u32): {:?}", expected_u32);
-
-    assert_eq!(
-        actual, expected,
-        "MASM from_account_id output should match Rust EthEmbeddedAccountId::to_elements()"
-    );
-
-    Ok(())
-}
-
 /// Tests that two messages with identical destinations but different metadata_hashes produce
 /// different LET roots.
 ///
